Use migrate to organize and future-proof database migrations
Move the database initialization step to a migration. Migrations are
SQL files stored in a subdirectory. They are embedded into the
resulting binary.

Structuring migrations using the migrate go pkg makes future changes
to the database structure possible. They can just be added to the
migrations subdirectory and will be applied automatically by the
application on start.

This change will not break any running deployments of
timew-sync-server. The only database migration currently is the
already existing database initialization, which always has been
idempotent and won't modify or delete data.